test(halo2): move `SimpleCircuit` test codes from `circuit.rs` to `circuits/simple_circuit.rs`
This is because we want to add a circuit test within the associated circuit code.
